feat: match against directories in $XDG_DATA_DIRS to check for backgrounds
At the moment, we check if a `source` is located under
`/usr/share/backgrounds/` and only then do we recurse into it for more
potential background images.

The `/usr/share` directory doesn't exist on _all_ Linux distributions.
On NixOS, the `/usr` directory doesn't exist. What one would usually
find in `/usr/share` is found in `/run/current-system/sw/share` in
NixOS.

But, the paths `/usr/share` and `/run/current-system/sw/share` are
included in the environment variable $XDG_DATA_DIRS. This approach of
using the value of $XDG_DATA_DIRS is beneficial for two reasons:

1. We avoid hard-coding distribution-specific paths like
   `/run/current-system/sw`.
2. Properly initializing the $XDG_DATA_DIRS variable is the
   distribution's responsibility. Not setting it up correctly with
   system's data directories **should never happen**.